Create Main Bot Code
Code for a bull-bot that charges in one direction after detecting an infrared wave, akin to an actual bull seeing a red cape held up by a matador. Our robot, however, is a little more timid, or perhaps even playful as an onboard ultrasonic sensor will prevent the bot from colliding head-on with an object. The bull is moved by two dc motors and the horns are maneuvered by two finite servos. Controlling the bull is done via an Arduino and the respective code is written in C++.